feat: Settings component for user settings management
